backend/requirements/docker.txt
# Minimal requirements for Docker builds
# Only essential packages to get the backend running

# FastAPI and ASGI
fastapi==0.105.0
uvicorn[standard]==0.24.0

# AI & ML - Essential only
google-generativeai

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Utilities (use flexible versions to avoid conflicts)
pydantic[email]>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2